+++ src/commonMain/kotlin/com/hoc081098/flowext/internal/ClosedException.kt
@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2021-2023 Petrus Nguyễn Thái Học
+ * Copyright (c) 2021-2024 Petrus Nguyễn Thái Học
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -24,17 +24,22 @@
 
 package com.hoc081098.flowext.internal
 
-import kotlinx.coroutines.flow.FlowCollector
+// Reference: https://github.com/Kotlin/kotlinx.coroutines/blob/8c516f5ab1fcc39629d2838489598135eedd7b80/kotlinx-coroutines-core/common/src/flow/internal/FlowExceptions.common.kt
+// Change: We don't inherit from `kotlinx.coroutines.CancellationException`.
 
 /**
- * This exception is thrown when operator need no more elements from the flow.
- * This exception should never escape outside of operator's implementation.
- * This exception can be safely ignored by non-terminal flow operator if and only if it was caught by its owner
- * (see usages of [checkOwnership]).
+ * This exception is thrown when an operator needs no more elements from the flow.
+ *
+ * The operator should never allow this exception to be thrown past its own boundary.
+ * This exception can be safely ignored by non-terminal flow operator
+ * if and only if it was caught by its owner (see usages of [checkOwnership]).
+ *
+ * Therefore, the [owner] parameter must be unique for every invocation of every operator.
  */
-internal class ClosedException(val owner: FlowCollector<*>) :
-  Exception("Flow was aborted, no more elements needed")
+internal expect class ClosedException(owner: Any) : Exception {
+  val owner: Any
+}
 
-internal fun ClosedException.checkOwnership(owner: FlowCollector<*>) {
+internal fun ClosedException.checkOwnership(owner: Any) {
   if (this.owner !== owner) throw this
 }